backend/requirements.txt
fastapi
uvicorn
pymongo
scikit-learn
numpy
requests
apscheduler
pydantic
psycopg2-binary